requirements/base.txt

# Django
django~=3.0.5

# Django Auth
django-allauth~=0.41.0

# Django Configuration
django-environ~=0.4.5

# Database Drivers
psycopg2-binary~=2.8.5
psycopg2~=2.8.5


# Django Forms
django-crispy-forms~=1.9.0

# Password Storage
argon2-cffi~=19.2.0

# Images
pillow~=7.1.2

# Restful Api
coreapi~=2.3.3
django-rest-auth~=0.9.5
djangorestframework-jwt~=1.11.0
djangorestframework~=3.11.0
drf-nested-routers~=0.91
drf-yasg~=1.17.1